Isolate collect_bytes()
Internal-tag: [#51803]
